Refactor model for export
