+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "@oliverbo/gg-common",
-    "version": "1.0.0",
+    "version": "1.1.0",
     "description": "Common functionality for glamglare site components",
     "main": "dist/index.js",
     "scripts": {
