+++ src/index.ts
@@ -6,8 +6,7 @@ export * as NotFoundError from "./error/NotFoundError";
 export * as ApiError from "./error/NotFoundError";
 
 // export model interfaces
-export * as Id from "./model/Id";
-export * as Artist from "./model/Artist";
-export * as ImageInfo from "./model/ImageInfo";
-export * as Post from "./model/Post";
-export * as Song from "./model/Song";
+// export * as Id from "./model/Id";
+// export * as Artist from "./model/Artist";
+// export * as xxx from "./model/Album";
+export * as model from "./model";
