--- src/model/Id.ts
@@ -1,4 +1,0 @@
-export interface Id {
-    id?: string;
-    slug?: string;
-}
