--- src/model/ImageInfo.ts
@@ -1,6 +1,0 @@
-
-export interface ImageInfo {
-    url: string;
-    width: number;
-    height: number;
-}
