--- src/model/Song.ts
@@ -1,7 +1,0 @@
-import { Artist } from './Artist';
-
-// Song
-export interface Song {
-    artist: Artist;
-    title: string;
-}
